backend/backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
joblib==1.5.3
lxml==6.0.2
numpy==2.4.3
pandas==3.0.1
pycparser==3.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-whois==0.9.6
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
